split `generateSettingsList.js` into separate files
