+++ patches/patchesSettings/style.css
@@ -302,7 +302,7 @@
 .apply-button,
 .reset-button {
     padding: 10px 24px;
-    border-radius: 20px;
+    border-radius: 25px;
     font-size: 14px;
     font-weight: 500;
     cursor: pointer;
